% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataprov.R
\docType{package}
\name{dataprov}
\alias{dataprov}
\alias{dataprov-package}
\title{Easy Data Provenance Tracking}
\usage{
dataprov()
}
\value{
A dataprov object, which is a data frame with the following fields for each entry (row):
 \item{timestamp}{A timestamp of the entry; see \code{\link{DateTimeClasses}}}
 \item{caller}{Name of the function that added this entry}
 \item{message}{Message describing the operation}
 \item{digest}{MD5 hash of the parent object as entry was made}
}
\description{
*Data provenance* broadly refers to a description of the origins of a piece
of data and the process by which it arrived in a database or analysis.
This package implements a simple system for tracking operations performed
on any R object.

Constructor.
}
\keyword{internal}

