% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getProvenance.R
\name{listProvenances}
\alias{listProvenances}
\title{List provenances.}
\usage{
listProvenances(env = .GlobalEnv)
}
\arguments{
\item{env}{Environment to look in for provenances (optional)}
}
\value{
A data frame listing any provences found
}
\description{
List provenances.
}

