% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataprov.R
\name{print.provenance}
\alias{print.provenance}
\title{Print a 'dataprov' class object.}
\usage{
\method{print}{provenance}(x, ...)
}
\arguments{
\item{x}{A \code{\link{dataprov}} object}

\item{...}{Other parameters passed to cat}
}
\description{
Print a 'dataprov' class object.
}
\details{
Prints a one-line summary of the object
}
\keyword{internal}

