% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSetProvenance.R
\name{provenance<-}
\alias{provenance<-}
\title{Replace entire provenance.}
\usage{
provenance(x) <- value
}
\arguments{
\item{x}{An object.}

\item{p}{A data frame that includes the required provenance fields (see \code{\link{provenance}}):}
}
\value{
x, with provenance set to contents of p
}
\description{
Replace entire provenance.
}
\examples{
d1 <- updateProvenance(cars, "first d1 message")
d2 <- updateProvenance(cars, "first d2 message")
provenance(d1) <- provenance(d2)
provenance(d1)
}

