% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getProvenance.R
\name{provenance}
\alias{provenance}
\title{Get provenance.}
\usage{
provenance(x, n = NULL, env = .GlobalEnv)
}
\arguments{
\item{x}{An object}

\item{n}{The line of the provenance to return as a list (optional)}
}
\value{
A provenance object, as a data frame, with one row for each operation and four columns:
 \item{timestamp}{Date and time entry was added}
 \item{caller}{The function that added this entry, including its parameter values}
 \item{message}{Description of action(s) taken}
 \item{digest}{Hash of the data when this entry was added; see \code{\link{digest}}}
}
\description{
Get provenance information of an object.
}
\examples{
d <- updateProvenance(cars, "first message")
d <- updateProvenance(d, "second message")
provenance(d)
}

