% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setProvenance.R
\name{replaceProvenance}
\alias{replaceProvenance}
\title{Replace entire provenance.}
\usage{
replaceProvenance(x, newp, env = .GlobalEnv)
}
\arguments{
\item{x}{An object.}

\item{p}{A data frame that includes the required provenance fields (see \code{\link{provenance}}):}
}
\value{
x, with provenance set to contents of p
}
\description{
Replace entire provenance.
}
\keyword{internal}

