% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updateProvenance.R
\name{updateProvenance}
\alias{updateProvenance}
\title{Add provenance information to an object}
\usage{
updateProvenance(x, message, caller = NULL)
}
\arguments{
\item{x}{An R object}

\item{message}{A string (message to be added to the provenance)}

\item{caller}{The calling function (automatically filled in if NULL)}
}
\value{
The original object, with an updated provenance containing:
 \item{timestamp}{Date and time entry was added}
 \item{caller}{The function that added this entry, including its parameter values}
 \item{message}{Description of action(s) taken}
 \item{digest}{Hash of the data when this entry was added; see \code{\link{digest}}}
This function logs information from the caller to a 'provenance' data structure.
}
\description{
It's important to track data provenance, the steps taken to produce a particular
dataset.
}
\examples{
d <- updateProvenance(cars, "first message")
d <- updateProvenance(d, "second message")
provenance(d)
}

