% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSetProvenance.R
\name{replaceProvenance}
\alias{replaceProvenance}
\title{Replace entire provenance.}
\usage{
replaceProvenance(x, p)
}
\arguments{
\item{x}{An object.}

\item{p}{A data frame that includes the required provenance fields (see \code{\link{provenance}}):}
}
\value{
x, with provenance set to contents of p
}
\description{
Replace entire provenance.
}
\examples{
d1 <- updateProvenance(cars, "first d1 message")
d2 <- updateProvenance(cars, "first d2 message")
d1 <- replaceProvenance(d1, provenance(d2))
provenance(d)
}

